Add config profile to install Chrome Extensions
It appears that as of version 59, Google Chrome will not automatically
install extension added to the "External Extensions" folder in
/Library/Application\ Support/Google/Chrome/. It therefore appears that
installing Chrome extensions by package is deprecated.*

Instead, extensions should be installed by profile. This has the
advantage of automatic forced installation and the inability of users to
uninstall these extensions. Initial tests appear to show that multiple
profiles will not work; all extensions should be included in a single
profile (this profile includes the uBlock Origin, HTTPS Everywhere, and
LastPass extensions).

*Installation will still work, but requires the user to approve each
extension.