Create first launchd LaunchAgent
Adds the files for a launchd LaunchAgent to ensure OneDrive is open in
the background (to ensure syncing is enabled). These include the
LaunchAgent (.plist file) and the job (shell script). This is the first
launchd LauchDaemon/LaunchAgent in this repository. More will (likely)
follow.

Going forward, these component files will need to be added to a package.
One way to do this would be a script which generates the necessary files
and builds a package using munkipkg.