Add printer nopkgs
Adds the first nopkgs (no package) munki scripts. These are munki
pkginfo files that are associated with no pkg or app. Instead, they are
used to run postinstall scripts, either on-demand or based on an install
check script.

These initial nopkgs are focused on printers. Others should be added in
the future.